Extract `Storage` class
