+++ src/main/java/Duke.java
@@ -1,10 +1,6 @@
 import java.util.Scanner;
 import java.util.List;
-import java.util.ArrayList;
 import java.lang.StringBuilder;
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
 
 public class Duke {
     private static String WELCOME_TEXT = "Hey there I'm Duke!\n" + "How can I help you today?";
@@ -12,118 +8,18 @@ public class Duke {
 
     private Scanner in;
     private List<Task> tasks;
-    private File tasksFile;
+    private Storage storage;
     private boolean shouldExit;
 
     public enum Command {
         BYE, LIST, DONE, DELETE, TODO, DEADLINE, EVENT,
     }
 
-    Duke(Scanner in, String filePath) throws IOException {
+    Duke(Scanner in, String filePath) throws Exception {
         this.in = in;
         this.shouldExit = false;
-
-        File dataDirectory = new File("data");
-        if (!dataDirectory.exists()) {
-            dataDirectory.mkdir();
-        }
-        File tasksFile = new File("data/" + filePath);
-        if (!tasksFile.exists()) {
-            tasksFile.createNewFile();
-        }
-        this.tasksFile = tasksFile;
-        try {
-            this.tasks = this.loadTasks();
-        } catch (Exception e) {
-            this.tasks = new ArrayList<>();
-        }
-    }
-
-    private String escapeString(String str) {
-        return str.replace("|", "||");
-    }
-
-    private String unescapeString(String str) {
-        return str.replace("||", "|");
-    }
-
-    private List<Task> loadTasks() throws Exception {
-        List<Task> tasks = new ArrayList<>();
-        Scanner s = new Scanner(this.tasksFile);
-        try {
-            while (s.hasNext()) {
-                String line = s.nextLine();
-                String[] components = line.split(" \\| ");
-                if (components.length < 3) {
-                    throw new Exception("Invalid format");
-                }
-                boolean isCompleted = components[1] == "1" ? true : false;
-                String description = unescapeString(components[2]);
-                switch (components[0]) {
-                    case "T": {
-                        tasks.add(new Todo(description, isCompleted));
-                        break;
-                    }
-                    case "E": {
-                        if (components.length != 4) {
-                            throw new Exception("Invalid format");
-                        }
-                        tasks.add(new Event(description, DateTime.parse(components[3]), isCompleted));
-                        break;
-                    }
-                    case "D": {
-                        if (components.length != 4) {
-                            throw new Exception("Invalid format");
-                        }
-                        tasks.add(new Deadline(description, DateTime.parse(components[3]), isCompleted));
-                        break;
-                    }
-                    default: {
-                        throw new Exception("Invalid format");
-                    }
-                }
-            }
-        } finally {
-            s.close();
-        }
-
-        return tasks;
-    }
-
-    private void saveTasks() throws IOException {
-        FileWriter writer = new FileWriter(this.tasksFile);
-
-        for (Task task : this.tasks) {
-            this.writeTask(writer, task);
-        }
-
-        writer.close();
-    }
-
-    private void writeTask(FileWriter writer, Task task) throws IOException {
-        if (task instanceof Todo) {
-            writer.write("T | ");
-            writer.write(task.isCompleted ? "1" : "0");
-            writer.write(" | ");
-            writer.write(escapeString(task.description));
-            writer.write(System.lineSeparator());
-        } else if (task instanceof Event) {
-            writer.write("E | ");
-            writer.write(task.isCompleted ? "1" : "0");
-            writer.write(" | ");
-            writer.write(escapeString(task.description));
-            writer.write(" | ");
-            writer.write(DateTime.stringify(((Event) task).time));
-            writer.write(System.lineSeparator());
-        } else if (task instanceof Deadline) {
-            writer.write("D | ");
-            writer.write(task.isCompleted ? "1" : "0");
-            writer.write(" | ");
-            writer.write(escapeString(task.description));
-            writer.write(" | ");
-            writer.write(DateTime.stringify(((Deadline) task).time));
-            writer.write(System.lineSeparator());
-        }
+        this.storage = new Storage(filePath);
+        this.tasks = storage.loadTasks();
     }
 
     private void greet() {
@@ -196,7 +92,7 @@ public class Duke {
                     throw new Exception("There is no task with the following number: " + arguments);
                 }
 
-                this.saveTasks();
+                this.storage.saveTasks(this.tasks);
                 printMessage("Marking task as completed:\n    " + task.toString());
                 break;
             }
@@ -215,7 +111,7 @@ public class Duke {
                     throw new Exception("There is no task with the following number: " + arguments);
                 }
 
-                this.saveTasks();
+                this.storage.saveTasks(this.tasks);
                 printMessage("Removed the following task:\n    " + task.toString() + "\n" + "You now have "
                         + this.tasks.size() + " tasks in your list.");
                 break;
@@ -224,7 +120,7 @@ public class Duke {
                 Task todo = Todo.fromInput(arguments);
                 this.tasks.add(todo);
 
-                this.saveTasks();
+                this.storage.saveTasks(this.tasks);
                 this.printTaskAddedMessage(todo);
                 break;
             }
@@ -232,7 +128,7 @@ public class Duke {
                 Task deadline = Deadline.fromInput(arguments);
                 this.tasks.add(deadline);
 
-                this.saveTasks();
+                this.storage.saveTasks(this.tasks);
                 this.printTaskAddedMessage(deadline);
                 break;
             }
@@ -240,7 +136,7 @@ public class Duke {
                 Task event = Event.fromInput(arguments);
                 this.tasks.add(event);
 
-                this.saveTasks();
+                this.storage.saveTasks(this.tasks);
                 this.printTaskAddedMessage(event);
                 break;
             }
@@ -278,9 +174,8 @@ public class Duke {
         while (!duke.shouldExit()) {
             try {
                 duke.handleInput();
-            }
-            // TODO: custom Duke exceptions?
-            catch (Exception e) {
+            } catch (Exception e) {
+                // TODO: custom Duke exceptions?
                 printMessage("Error: " + e.getMessage());
             }
         }
